update graphics.js and app.js
moved map data to App.js
